refactor(infrastructure): rename logings module
